feat: change asset page structure
