+++ src/renderer/entities/asset/index.ts
@@ -1,1 +1,2 @@
 export * from './ui';
+export { assetsModel } from './model/assets-model';
