+++ src/renderer/features/assets/AssetsChainView/ui/AssetCard/AssetCard.tsx
