+++ src/renderer/features/assets/AssetsChainView/ui/NetworkAssets/NetworkAssets.test.tsx
