+++ src/renderer/features/assets/AssetsChainView/ui/NetworkFiatBalance/NetworkFiatBalance.tsx
