+++ src/renderer/features/assets/index.ts
@@ -1,1 +1,2 @@
 export { AssetRouteGuard } from './AssetRouteGuard';
+export { AssetsChainView } from './AssetsChainView/ui/AssetsChainView';
