+++ src/renderer/pages/Assets/Assets/components/index.ts
@@ -1,4 +1,1 @@
-export { AssetsList } from './AssetsList/AssetsList';
 export { AssetsFilters } from './AssetsFilter/AssetsFilters';
-export { NetworkAssets } from './NetworkAssets/NetworkAssets';
-export { AssetCard } from './AssetCard/AssetCard';
