+++ src/renderer/pages/Assets/Assets/model/__tests__/assets-view-model.test.ts
@@ -1,7 +1,8 @@
 import { fork, allSettled } from 'effector';
 
-import { assetsModel } from '../assets-model';
 import { localStorageService } from '@shared/api/local-storage';
+import { assetsModel } from '@entities/asset';
+import { assetsViewModel } from '../assets-view-model';
 
 describe('pages/Assets/Assets/model/assets-model', () => {
   beforeEach(() => {
@@ -13,7 +14,7 @@ describe('pages/Assets/Assets/model/assets-model', () => {
 
     const scope = fork();
 
-    await allSettled(assetsModel.events.assetsStarted, { scope });
+    await allSettled(assetsViewModel.events.assetsStarted, { scope });
 
     expect(scope.getState(assetsModel.$hideZeroBalances)).toEqual(false);
   });
@@ -23,7 +24,7 @@ describe('pages/Assets/Assets/model/assets-model', () => {
 
     const scope = fork();
 
-    await allSettled(assetsModel.events.assetsStarted, { scope });
+    await allSettled(assetsViewModel.events.assetsStarted, { scope });
 
     expect(scope.getState(assetsModel.$hideZeroBalances)).toEqual(true);
   });
@@ -36,18 +37,8 @@ describe('pages/Assets/Assets/model/assets-model', () => {
       values: new Map().set(assetsModel.$hideZeroBalances, false),
     });
 
-    await allSettled(assetsModel.events.hideZeroBalancesChanged, { scope, params: newValue });
+    await allSettled(assetsViewModel.events.hideZeroBalancesChanged, { scope, params: newValue });
 
     expect(scope.getState(assetsModel.$hideZeroBalances)).toEqual(newValue);
-  });
-
-  test('should update $query on change', async () => {
-    const scope = fork({
-      values: new Map().set(assetsModel.$query, 'hello'),
-    });
-
-    await allSettled(assetsModel.events.queryChanged, { scope, params: 'world' });
-
-    expect(scope.getState(assetsModel.$query)).toEqual('world');
   });
 });
